python_ml/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
dnspython==2.7.0
faiss-cpu==1.9.0.post1
fastapi==0.115.5
filelock==3.16.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
motor==3.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
